Better checks traceroute output
Traceroute doesn't always manage to reach a remote host for a variety
of reasons: network down, unreachable host, etc. This adds a check by
pinging the host over the expected interface, if the ping fails then
no result is returned.

Later, when we would ping the first-hop we check that such a hop
exists, if one was not returned then no ping is done.

The re-init timer was lowered from 1h to 15m to mitigate temporary
network errors that would remove the first-hop data.