Switch to multiplier modifier for lost latency
This sets the latency of lost packets to be equal to the maximum
latency seen so far multiplied by a flag value.

Using a $MAX * $mod for setting the lost packet latency allows
callers to return to a single pass computation, instead of running
over data once to determine the max value, then pass it to the
program.